requirements.txt
mysqlclient
Django
Pillow
django-imagekit
